how to add two big intiger using string in c language
With various data type we can add only those two numbers those are in the range of declared data type. But we want to add two numbers those are out of this range , using string is a way that can be helpful. So I have tried to make a program that can add two big number.